Reverting 5882. It broke LoginPromptTest.TestBasicAuth and .TestTwoAuths. Review URL: http://codereview.chromium.org/11399
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@5922 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/tab_contents.h
@@ -393,8 +393,8 @@ class TabContents : public PageNavigator,
   void RemoveInfoBar(InfoBarDelegate* delegate);
   
   // Enumeration and access functions.
-  int infobar_delegate_count() const { return infobar_delegates_.size(); }
-  InfoBarDelegate* GetInfoBarDelegateAt(int index) {
+  size_t infobar_delegate_count() const { return infobar_delegates_.size(); }
+  InfoBarDelegate* GetInfoBarDelegateAt(size_t index) {
     return infobar_delegates_.at(index);
   }
